+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 - add missing service identity accounts (artifactregistry, composer) in project module
 - new "Cloud Storage to Bigquery with Cloud Dataflow with least privileges" example
 - support service dependencies for crypto key bindings in project module
+- refactor project module in multiple files
+- add support for per-file option overrides to tfdoc
 
 ## [12.0.0] - 2022-01-11